requirements.txt
python-pptx
scipy